api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--no-binary psycopg2

apipkg==1.5               # via execnet
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
beautifulsoup4==4.7.1     # via django-livereload-server
django-livereload-server==0.3.2
django-revproxy==0.9.15
django==1.11.20
djangorestframework==3.9.4
execnet==1.6.0            # via pytest-xdist
more-itertools==7.0.0     # via pytest
pluggy==0.11.0            # via pytest
psycopg2==2.8.2
py==1.8.0                 # via pytest
pytest-django==3.4.8
pytest-forked==1.0.2      # via pytest-xdist
pytest-xdist==1.28.0
pytest==4.5.0
pytz==2019.1              # via django
six==1.12.0               # via django-livereload-server, pytest, pytest-xdist
soupsieve==1.9.1          # via beautifulsoup4
tornado==6.0.2            # via django-livereload-server
urllib3==1.25.2           # via django-revproxy
wcwidth==0.1.7            # via pytest
